+++ lib/models/task.dart
@@ -26,6 +26,10 @@ class Task {
   /// This [remindMe] is used to check if the task is should be reminded.
   bool remindMe;
 
+  /// This [addDueDate] is used to check if the due date for the reminder of
+  /// the task has been set.
+  bool addDueDate;
+
   Task({
     required this.id,
     required this.step,
@@ -35,6 +39,7 @@ class Task {
     this.myDay = false,
     this.isSelected = false,
     this.remindMe = false,
+    this.addDueDate = false,
   });
 
   Task copyWith({
@@ -46,6 +51,7 @@ class Task {
     bool? myDay,
     bool? isSelected,
     bool? remindMe,
+    bool? addDueDate,
   }) {
     return Task(
       id: id ?? this.id,
@@ -56,6 +62,7 @@ class Task {
       myDay: myDay ?? this.myDay,
       isSelected: isSelected ?? this.isSelected,
       remindMe: remindMe ?? this.remindMe,
+      addDueDate: addDueDate ?? this.addDueDate,
     );
   }
 }
